requirements.txt
click>=8.1.0
Pillow>=10.0.0
tqdm>=4.65.0
